切换光标为世界坐标系，接口与渲染全面升级
本次提交将光标位置的表示方式从体素索引（int[3]）统一切换为世界坐标（double[3]），涉及 AbstractInteractiveService、MedicalVizService、AppState 及各类渲染策略、交互 Handler 等接口和实现。RenderParams 结构体同步调整，所有交互与渲染逻辑均以世界坐标为核心，提升了多视图、多变换场景下的交互一致性和精度，为后续支持任意方向切片等功能奠定基础。相关注释与文档已同步更新。

+++ MVVCVTK/AppTypes.h
@@ -110,7 +110,7 @@ namespace VisFlags {
 
 // --- 渲染参数结构体（Strategy 的唯一输入，不含 VTK 指针）---
 struct RenderParams {
-    std::array<int, 3>     cursor = { 0, 0, 0 };
+    std::array<double, 3>  cursor = { 0, 0, 0 };
     std::vector<TFNode>    tfNodes;
     double                 scalarRange[2] = { 0.0, 255.0 };
     MaterialParams         material;
